Move GeoTiffInfo to geotrellis.raster.io.geotiff.reader
Moving out of GeoTiffReader class. This is a often used class in client code and it makes little sense to tuck it away in a reader.
