+++ raster/src/test/scala/geotrellis/raster/io/geotiff/GeoTiffMultibandTileSpec.scala
@@ -797,7 +797,7 @@ class GeoTiffMultibandTileSpec extends FunSpec
   describe("GeoTiffMultibandTile streaming read") {
     it("reads over-buffered windows"){
       val path = geoTiffPath("3bands/int32/3bands-striped-pixel.tif")
-      val info = GeoTiffReader.readGeoTiffInfo(
+      val info = GeoTiffInfo.read(
         ByteBuffer.wrap(Filesystem.slurp(path)),
         streaming = true, withOverviews = false,
         byteReaderExternal = None
