+++ s3-spark/src/main/scala/geotrellis/spark/store/s3/S3GeoTiffInfoReader.scala
@@ -17,7 +17,7 @@
 package geotrellis.spark.store.s3
 
 import geotrellis.raster.io.geotiff.reader.GeoTiffReader
-import geotrellis.raster.io.geotiff.reader.GeoTiffReader.GeoTiffInfo
+import geotrellis.raster.io.geotiff.reader.GeoTiffInfo
 import geotrellis.raster.io.geotiff.tags.TiffTags
 import geotrellis.store.s3._
 import geotrellis.store.s3.util.S3RangeReader
